Update _index.fr.md
Recently I entirely read Zig's website (why not) and discovered that it was available for french. How to say that the french translation is not horrible, as a french: I understand the most "important" (perhaps because I read the website first in english), but it's not feeling smooth to read.